backend/requirements.txt
asgiref==3.11.0
dj-database-url==3.1.0
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
